# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface SendOpenaiVoiceMessageBody {
  /** Base64-encoded audio data */
  audio: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "512x512": "512x512",
  "256x256": "256x256",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}

export interface JobRole {
  id: string;
  name: string;
  category: string;
  description: string;
}

export interface InterviewSession {
  id: number;
  jobRole: string;
  jobRoleName: string;
  conversationId: number;
  createdAt: string;
}

export interface InterviewSessionWithMessages {
  id: number;
  jobRole: string;
  jobRoleName: string;
  conversationId: number;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface CreateInterviewSessionBody {
  jobRole: string;
  jobRoleName: string;
}

export interface SendInterviewMessageBody {
  content: string;
}

export interface InterviewFeedback {
  sessionId: number;
  overallScore: number;
  strengths: string[];
  areasForImprovement: string[];
  summary: string;
}