# lib/api-zod/src/generated/types/interviewFeedback.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewFeedback {
  sessionId: number;
  overallScore: number;
  strengths: string[];
  areasForImprovement: string[];
  summary: string;
}