# lib/api-zod/src/generated/types/interviewSessionWithMessages.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OpenaiMessage } from "./openaiMessage";

export interface InterviewSessionWithMessages {
  id: number;
  jobRole: string;
  jobRoleName: string;
  conversationId: number;
  createdAt: Date;
  messages: OpenaiMessage[];
}